ユーザーを判定する機能を追加した

+++ app/Http/Controllers/BlogController.php
@@ -58,6 +58,12 @@ class BlogController extends Controller
      */
     public function destroy($id)
     {
+        $login_user_id = Auth::id();
+        $blog_user_id = Blog::find($id)->user_id;
+        if ($login_user_id !== $blog_user_id) {
+            abort(403);
+        }
+
         $blog = Blog::find($id);
         $blog->delete();
         return redirect()->route('blog.create');
@@ -68,6 +74,11 @@ class BlogController extends Controller
      */
     public function edit($id)
     {
+        $login_user_id = Auth::id();
+        $blog_user_id = Blog::find($id)->user_id;
+        if ($login_user_id !== $blog_user_id) {
+            abort(403);
+        }
         $blog = Blog::find($id);
         return view("edit")->with('blog', $blog);
     }
@@ -77,15 +88,21 @@ class BlogController extends Controller
      */
     public function update(Request $request, $id)
     {
+        $login_user_id = Auth::id();
+        $blog_user_id = Blog::find($id)->user_id;
+        if ($login_user_id !== $blog_user_id) {
+            abort(403);
+        }
+        
         $request->validate([
             'title' => ['required'],
             'body' => ['required', 'min:10'],
         ]);
-
         Blog::find($id)->update([
             'title' => $request->title,
             'body' => $request->body
         ]);
         return redirect()->route('top');
+
     }
 }
